+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.8.11 (2020-01-dd)
 
 - fix(web): Fix parsing of event search parameters (including limit)
+- fix(db): Implicitly convert all tags to lowercase
 
 ## v0.8.10 (2020-01-23)